+++ Interface/ImguiParticleLogger.h
@@ -17,7 +17,7 @@ private:
 	std::string LogToText(const std::string& separator = "\t", FILE* file = nullptr);
 	bool enableLogging;
 	std::string facetNumInput = "";
-	int facetNum = -1;
+	size_t facetNum = -1;
 	std::string maxRecInput = "10000";
 	size_t maxRec = 10000;
 	bool isRunning = false;
